refactor(digest): rename hashx package to digest
Descriptive domain noun (SHA/HMAC digests) beats the stdlib-collision x
suffix; no stdlib 'hash' clash for the base concept here. No external
consumers. Test-local var 'digest' renamed to 'raw' to avoid shadowing.
